Refactor models Directory and Archive to be Nameable
